Created timeframe module
