+++ populate_db.py
@@ -1,14 +1,9 @@
 import sqlite3
 from account_data_retriever import AccountDataRetriever
 from data_retriever import DataRetriever
-from enum import Enum
 from datetime import date, timedelta
 
-class TimeFrame(Enum):
-  Day = "1Day"
-  Hour = "1Hour"
-  Minute = "1Min"
-  Sec = "1Sec"
+
 
 ADR = AccountDataRetriever()
